informatica_api/app/main.py
import os

from fastapi import FastAPI, HTTPException

from pydantic import BaseModel

from langchain_core.messages import HumanMessage

from dotenv import load_dotenv

from app.agent import agent


load_dotenv()


app = FastAPI(

    title="AlgoRhythm Agent API",

    description="A content creator agent powered by LangGraph",

    version="1.0.0"

)


class ChatRequest(BaseModel):

    message: str


class ChatResponse(BaseModel):

    response: str


@app.get("/health")

async def health_check():

    return {"status": "ok"}


@app.post("/chat", response_model=ChatResponse)

async def chat(request: ChatRequest):

    try:

        result = agent.invoke({

            "messages": [HumanMessage(content=request.message)]

        })

        return ChatResponse(response=result["messages"][-1].content)

    except Exception as e:

        raise HTTPException(status_code=500, detail=str(e))